# Conflicts:
#	src/app/components/diary/diary-card/diary-card.component.spec.ts
#	src/app/diary/diary-card/diary-card.component.ts
#	src/app/diary/diary-show/diary-show.component.html

+++ src/app/components/admin/admin-show-user-list.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {AdminUserService} from '../../services/admin/admin-user.service';
 import {ActivatedRoute} from '@angular/router';
 import {User} from '../../models/User';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 
 @Component({
   selector: 'app-admin-show-user-list',
@@ -10,13 +11,16 @@ import {User} from '../../models/User';
 })
 export class AdminShowUserListComponent implements OnInit {
   userList: User[] = [];
-  page: number = 0;
+  page = 0;
   pages: Array<number>;
   user: User;
   changeText: boolean;
+  pickUpUser: User;
+  editForm: FormGroup;
 
   constructor(private usersService: AdminUserService,
-              private activatedRoute: ActivatedRoute) {
+              private activatedRoute: ActivatedRoute,
+              private formBuilder: FormBuilder) {
     this.changeText = false;
   }
 
@@ -30,6 +34,14 @@ export class AdminShowUserListComponent implements OnInit {
   ngOnInit(): void {
     // this.usersService.showListUser().subscribe(result => this.userList = result, error => console.log(error));
     this.getAllUserList();
+
+    this.editForm = this.formBuilder.group({
+        username: ['', [Validators.minLength(3), Validators.required]],
+        email: ['', [Validators.required, Validators.email]],
+        roles: [''],
+      }
+    );
+    this.usersService.shouldRefresh.subscribe(res => this.getAllUserList());
   }
 
   private getAllUserList(): void {
@@ -46,15 +58,60 @@ export class AdminShowUserListComponent implements OnInit {
   }
 
   deleteUser(id: number): void {
-    this.usersService.deleteUser(id).subscribe(() => this.getAllUserList(), error => console.log(error));
+    if (confirm('Do you really want to delete this user ?')) {
+      this.usersService.deleteUser(id).subscribe(() => this.getAllUserList(), error => console.log(error));
+    }
   }
 
   blockUser(i: number): void {
-    this.usersService.blockUser(i).subscribe(() => {console.log('block ok'); this.getAllUserList(); }, error => console.log(error));
+    this.usersService.blockUser(i).subscribe(() => {
+      console.log('block ok');
+      this.getAllUserList();
+    }, error => console.log(error));
   }
 
 
   activeUser(i: number): void {
-    this.usersService.activeUser(i).subscribe(() => {console.log('active ok'); this.getAllUserList(); }, error => console.log(error));
+    this.usersService.activeUser(i).subscribe(() => {
+      console.log('active ok');
+      this.getAllUserList();
+    }, error => console.log(error));
+  }
+
+  showDetailUser(id: string): void {
+    this.usersService.getUserById(+id).subscribe(result => {
+      this.pickUpUser = result;
+      console.log(this.pickUpUser);
+      this.editForm.patchValue(result);
+    });
+  }
+
+  changeInfor(closeModalRef: HTMLButtonElement): void {
+    console.log(this.editForm.get('roles').value);
+    const {value} = this.editForm;
+
+    const data = {
+      ...this.pickUpUser,
+      ...value
+    };
+
+    if (this.editForm.get('roles').value == 1) {
+      data.roles = [{
+        id: '1',
+        name: 'ROLE_USER'
+      }];
+    } else {
+      console.log('admin');
+      data.roles = [{
+        id: '3',
+        name: 'ROLE_ADMIN'
+      }];
+    }
+    console.log(data);
+    this.usersService.editUser(data)
+      .subscribe(() => {
+        this.usersService.shouldRefresh.next();
+        alert('Cập nhật thành công');
+      }, error => console.log(error));
   }
 }
